rename :: get 붙어있는 함수들 fetch로 변경
