+++ Projects/Services/DataModule/Sources/School/UseCase/Impl/FetchSchoolListUseCaseImpl.swift
@@ -1,0 +1,17 @@
+import Combine
+import DataMappingModule
+import DomainModule
+import ErrorModule
+
+public struct FetchSchoolListUseCaseImpl: FetchSchoolListUseCase {
+    private let schoolRepository: any SchoolRepository
+
+    public init(schoolRepository: any SchoolRepository) {
+        self.schoolRepository = schoolRepository
+    }
+
+    public func execute() -> AnyPublisher<[SchoolEntity], DmsError> {
+        schoolRepository.fetchSchoolList()
+    }
+
+}
