+++ Projects/Services/DataModule/Sources/School/UseCase/Impl/FetchSchoolQuestionUseCaseImpl.swift
@@ -3,15 +3,15 @@ import DataMappingModule
 import DomainModule
 import ErrorModule
 
-public struct GetSchoolListUseCaseImpl: GetSchoolListUseCase {
+public struct FetchSchoolQuestionUseCaseImpl: FetchSchoolQuestionUseCase {
     private let schoolRepository: any SchoolRepository
 
     public init(schoolRepository: any SchoolRepository) {
         self.schoolRepository = schoolRepository
     }
 
-    public func execute() -> AnyPublisher<[SchoolEntity], DmsError> {
-        schoolRepository.fetchSchoolList()
+    public func execute(authCode: String) -> AnyPublisher<String, DmsError> {
+        schoolRepository.fetchSchoolQuestion(authCode: authCode)
     }
 
 }
