+++ Projects/Services/DomainModule/Sources/School/UseCases/GetSchoolQuestionUseCase.swift
@@ -2,6 +2,6 @@ import Combine
 import DataMappingModule
 import ErrorModule
 
-public protocol GetSchoolQuestionUseCase {
+public protocol FetchSchoolQuestionUseCase {
     func execute(authCode: String) -> AnyPublisher<String, DmsError>
 }
